# api/openapi.yaml
openapi: 3.0.3
info:
  title: Open API Kundenbeziehung - Swiss National Standard
  description: |
    Schweizer Standard für den branchenübergreifenden Austausch von Kundendaten zur Etablierung einer selbstbestimmten digitalen Kundenbeziehung.
    
    Diese API implementiert das vollständige Framework der Open API Kundenbeziehung mit dem universellen 10-Stufen-Referenzprozess und ermöglicht standardisierten Datenaustausch zwischen Finanzdienstleistern und anderen Ecosystemen.
    
    **Implementierte Module (Datenbausteine):**
    - Initialisierung (Step 1) - Service Discovery & Cookie Consent
    - Produktauswahl (Step 2) - Bedürfnisbefriedigung & Konfiguration  
    - Selbstdeklaration (Step 3) - FATCA/MIFID II Information
    - Basisdaten (Step 4) - Identität, Kontakt, Adresse
    - Erweiterte Daten (Step 5) - Risiko-/Potenzialermittlung
    - Identifikation (Step 6) - E-ID kompatibel, VideoIdent
    - Background Checks (Step 7) - KYC, AML, PEP, Sanctions
    - Vertragsabschluss (Step 8) - AGB-Akzeptanz
    - Signatur-Services (Step 9) - QES/AES/eSignatur
    - Metadaten-Management (Step 10) - Verarbeitung & Integration
    
    **Use Cases Implementiert:**
    - UC1: Kundenbeziehungseröffnung (Banking Account Opening)
    - UC2: Re-Identifikation (Cross-Provider Identity Reuse)
    - UC3: Altersverifikation (Privacy-preserving Age Verification)
    - UC4: CLM von EVV-Endkunden (External Wealth Management Lifecycle)
    
    **Sicherheitsstandards:**
...
    
...
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - mTLS: []
paths:
...
  /customer/check:
    post:
      summary: Kunden-Existenz prüfen
      description: Prüft ob ein Kunde bereits bei einer Institution identifiziert wurde (MVP Identifikation)
      operationId: checkCustomer
      tags:
        - Customer Data
      security:
        - mTLS: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCheckRequest'
            examples:
              basic_check:
                summary: Basis-Kundenprüfung
                value:
                  sharedCustomerHash: "sha256:abc123def456..."
                  basicData:
                    lastName: "Müller"
                    givenName: "Hans"
                    birthDate: "1985-03-15"
                    nationality: ["CH"]
      responses:
        '200':
          description: Kunde erfolgreich geprüft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /customer/fullRequest:
    post:
      summary: Vollständige Kundendaten anfordern
      description: Fordert das vollständige Kundendatenset an (erfordert gültigen Consent)
      operationId: requestFullCustomerData
      tags:
        - Customer Data
      security:
        - mTLS: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FullDataRequest'
      responses:
        '200':
          description: Kundendaten erfolgreich übertragen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullCustomerDataset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /customer/data:
    post:
      summary: Kundendaten mit Consent abrufen
      description: Ruft spezifische Kundendatenmodule basierend auf gewährtem Consent ab
      operationId: getCustomerData
      tags:
        - Customer Data
      security:
        - mTLS: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerDataRequest'
      responses:
        '200':
          description: Kundendaten erfolgreich abgerufen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

  # PORTFOLIO & EVV SERVICES
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-Token mit Consent-Claims für API-Zugriff (FAPI 2.0 compliant)
    DPoP:
      type: http
      scheme: DPoP
      description: DPoP (Demonstrating Proof-of-Possession) für sender-constrained tokens
    mTLS:
      type: mutualTLS
      description: Mutual TLS für höchste Sicherheitsanforderungen (FAPI 2.0 compliant)
    private_key_jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Private Key JWT client authentication (RFC 7523) - alternative zu mTLS
    RegistrationToken:
      type: http
      scheme: bearer
      description: Registration access token für dynamic client registration
  responses:
    BadRequest:
      description: Ungültige Anfrage
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "INVALID_REQUEST"
            message: "Missing required field: customerId"
            timestamp: "2024-01-15T10:30:00Z"
    Unauthorized:
      description: Nicht autorisiert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Invalid or missing authentication token"
            timestamp: "2024-01-15T10:30:00Z"
    Forbidden:
      description: Zugriff verweigert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "FORBIDDEN"
            message: "Insufficient permissions for requested operation"
            timestamp: "2024-01-15T10:30:00Z"
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Fehlercode
        message:
          type: string
          description: Fehlerbeschreibung
        timestamp:
          type: string
          format: date-time
          description: Zeitstempel des Fehlers
        details:
          type: object
          description: Zusätzliche Fehlerdetails
...
    CustomerCheckRequest:
      type: object
      required:
        - sharedCustomerHash
        - basicData
      properties:
        sharedCustomerHash:
          type: string
          description: SHA-256 Hash der Grunddaten für Matching
          example: "sha256:a1b2c3d4..."
        basicData:
          $ref: '#/components/schemas/BasicCustomerData'
    CustomerCheckResponse:
      type: object
      properties:
        match:
          type: boolean
          description: Ob ein Kunde gefunden wurde
        identificationDate:
          type: string
          format: date-time
          description: Datum der letzten Identifikation
        levelOfAssurance:
          type: string
          enum: [low, substantial, high]
          description: Sicherheitsniveau der Identifikation
        validUntil:
          type: string
          format: date-time
          description: Gültigkeit der Identifikation
    FullDataRequest:
      type: object
      required:
        - sharedCustomerHash
        - purpose
        - consentToken
      properties:
        sharedCustomerHash:
          type: string
          description: SHA-256 Hash des Kunden
        purpose:
          type: string
          enum: [accountOpening, creditAssessment, compliance]
        consentToken:
          type: string
          description: JWT-Token mit Consent-Nachweis
    FullCustomerDataset:
      type: object
      properties:
        basicData:
          $ref: '#/components/schemas/BasicCustomerData'
        contactInformation:
          $ref: '#/components/schemas/ContactInformation'
        addressData:
          $ref: '#/components/schemas/AddressData'
        identification:
          $ref: '#/components/schemas/IdentificationData'
        kycData:
          $ref: '#/components/schemas/KYCData'
        complianceData:
          $ref: '#/components/schemas/ComplianceData'
        riskProfile:
          $ref: '#/components/schemas/RiskProfile'
        metadata:
          $ref: '#/components/schemas/DataMetadata'
    BasicCustomerData:
      type: object
      required:
        - lastName
        - givenName
        - birthDate
        - nationality
      properties:
        lastName:
          type: string
          description: Nachname
          example: "Müller"
          maxLength: 100
        givenName:
          type: string
          description: Vorname(n)
          example: "Hans Peter"
          maxLength: 100
        middleName:
          type: string
          description: Weitere Vornamen
          maxLength: 100
        title:
          type: string
          description: Titel
          example: "Dr."
          maxLength: 50
        birthDate:
          type: string
          format: date
          description: Geburtsdatum
          example: "1985-03-15"
        birthPlace:
          type: string
          description: Geburtsort
          example: "Zürich"
          maxLength: 100
        nationality:
          type: array
          items:
            type: string
            pattern: '^[A-Z]{2}$'
          description: Nationalität(en) (ISO 3166-1)
          example: ["CH", "DE"]
          minItems: 1
          maxItems: 5
        gender:
          type: string
          enum: [male, female, other, unknown]
        maritalStatus:
          type: string
          enum: [single, married, divorced, widowed, registered_partnership]
        language:
          type: string
          description: Bevorzugte Sprache (ISO 639-1)
          example: "de"
          pattern: '^[a-z]{2}$'
    ContactInformation:
      type: object
      properties:
        primaryEmail:
          type: string
          format: email
          description: Primäre E-Mail-Adresse
          maxLength: 255
        secondaryEmail:
          type: string
          format: email
          description: Sekundäre E-Mail-Adresse
          maxLength: 255
        mobilePhone:
          type: string
          description: Mobiltelefonnummer (E.164 Format)
          example: "+41791234567"
          pattern: '^\+[1-9]\d{1,14}$'
        landlinePhone:
          type: string
          description: Festnetznummer
          pattern: '^\+[1-9]\d{1,14}$'
        preferredContactMethod:
          type: string
          enum: [email, sms, phone, postal, app]
        communicationLanguage:
          type: string
          description: Kommunikationssprache
          pattern: '^[a-z]{2}$'
        availabilityHours:
          type: string
          description: Verfügbarkeitszeiten
          example: "09:00-17:00"
    AddressData:
      type: object
      properties:
        residentialAddress:
          $ref: '#/components/schemas/Address'
        mailingAddress:
          $ref: '#/components/schemas/Address'
        businessAddress:
          $ref: '#/components/schemas/Address'
        previousAddresses:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Address'
              - type: object
                properties:
                  validFrom:
                    type: string
                    format: date
                  validTo:
                    type: string
                    format: date
    Address:
      type: object
      required:
        - street
        - postalCode
        - city
        - country
      properties:
        street:
          type: string
          description: Straße und Hausnummer
          example: "Bahnhofstrasse 1"
          maxLength: 255
        addressLine2:
          type: string
          description: Adresszusatz
          maxLength: 255
        postalCode:
          type: string
          description: Postleitzahl
          example: "8001"
          maxLength: 20
        city:
          type: string
          description: Ort
          example: "Zürich"
          maxLength: 100
        region:
          type: string
          description: Kanton/Region
          example: "ZH"
          maxLength: 100
        country:
          type: string
          description: Land (ISO 3166-1)
          example: "CH"
          pattern: '^[A-Z]{2}$'
        addressType:
          type: string
          enum: [residential, mailing, business, temporary]
    IdentificationData:
      type: object
      properties:
        identificationMethod:
          type: string
          enum: [video_identification, online_identification, personal_appearance, eid]
          description: Identifikationsmethode
        documentType:
          type: string
          enum: [passport, id_card, driving_license, eid]
          description: Dokumententyp
        documentNumber:
          type: string
          description: Dokumentennummer
          maxLength: 50
        issuingAuthority:
          type: string
          description: Ausstellende Behörde
          maxLength: 255
        issueDate:
          type: string
          format: date
          description: Ausstellungsdatum
        expiryDate:
          type: string
          format: date
          description: Ablaufdatum
        issuingCountry:
          type: string
          description: Ausstellungsland (ISO 3166-1)
          pattern: '^[A-Z]{2}$'
        levelOfAssurance:
          type: string
          enum: [low, substantial, high]
          description: Sicherheitsniveau (eIDAS-konform)
        verificationDate:
          type: string
          format: date-time
          description: Verifikationsdatum
        verificationMethod:
          type: string
          description: Verifikationsmethode
          maxLength: 100
        documentImageUrl:
          type: string
          format: uri
          description: Signierter Link zum Dokumentenscan (10min gültig)
        biometricData:
          type: object
          description: Biometrische Daten (wenn verfügbar)
          properties:
            faceImage:
              type: string
              description: Gesichtsbild (Base64)
            livenessScore:
              type: number
              minimum: 0
              maximum: 1
              description: Liveness-Check Score (0-1)
        nfcData:
          type: object
          description: NFC-Daten von eID
          properties:
            available:
              type: boolean
            verified:
              type: boolean
    KYCData:
      type: object
      properties:
        occupation:
          type: string
          description: Beruf
          maxLength: 255
        employer:
          type: string
          description: Arbeitgeber
          maxLength: 255
        employmentType:
          type: string
          enum: [employed, self_employed, unemployed, retired, student]
        annualIncome:
          $ref: '#/components/schemas/MonetaryAmount'
        totalAssets:
          $ref: '#/components/schemas/MonetaryAmount'
        sourceOfFunds:
          type: string
          enum: [salary, business_income, inheritance, investment_returns, other]
        sourceOfFundsDescription:
          type: string
          description: Beschreibung der Mittelherkunft
          maxLength: 1000
        pepStatus:
          type: boolean
          description: Politisch exponierte Person
        pepDetails:
          type: string
          description: Details zur PEP-Stellung
          maxLength: 1000
        beneficialOwners:
          type: array
          items:
            $ref: '#/components/schemas/BeneficialOwner'
        businessPurpose:
          type: string
          description: Geschäftszweck (bei Firmenkunden)
          maxLength: 1000
    ComplianceData:
      type: object
      properties:
        fatcaStatus:
          type: string
          enum: [us_person, non_us_person, uncertain]
        fatcaClassification:
          type: string
          description: FATCA-Klassifikation
          maxLength: 100
        crsReportable:
          type: boolean
          description: CRS-Meldepflicht
        taxResidencies:
          type: array
          items:
            $ref: '#/components/schemas/TaxResidency'
        tinNumbers:
          type: array
          items:
            $ref: '#/components/schemas/TINNumber'
        sanctionsScreening:
          $ref: '#/components/schemas/ScreeningResult'
        amlRiskRating:
          type: string
          enum: [low, medium, high]
        lastDueDiligence:
          type: string
          format: date-time
        nextReviewDate:
          type: string
          format: date
    RiskProfile:
      type: object
      properties:
        investmentExperience:
          type: string
          enum: [none, limited, good, extensive]
        investmentKnowledge:
          type: string
          enum: [basic, intermediate, advanced, professional]
        riskTolerance:
          type: string
          enum: [conservative, moderate, aggressive]
        investmentHorizon:
          type: string
          enum: [short, medium, long]
        investmentObjectives:
          type: array
          items:
            type: string
            enum: [capital_preservation, income, growth, speculation]
        esgPreferences:
          type: object
          properties:
            considerEsgFactors:
              type: boolean
            esgImportance:
              type: string
              enum: [not_important, somewhat_important, very_important]
            exclusions:
              type: array
              items:
                type: string
...
    MonetaryAmount:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          minimum: 0
          description: Betrag
        currency:
          type: string
          description: Währung (ISO 4217)
          example: "CHF"
          pattern: '^[A-Z]{3}$'
    DataMetadata:
      type: object
      properties:
        originator:
          type: string
          description: Ursprungsystem/Institution
          maxLength: 255
        createdAt:
          type: string
          format: date-time
        lastUpdated:
          type: string
          format: date-time
        version:
          type: string
          description: Datenversion
          maxLength: 50
        dataClassification:
          type: string
          enum: [public, internal, confidential, restricted]
        retentionPolicy:
          type: string
          description: Aufbewahrungsrichtlinie
          maxLength: 255
        legalBasis:
          type: string
          description: Rechtsgrundlage der Verarbeitung
          maxLength: 255
        qualityScore:
          type: number
          minimum: 0
          maximum: 1
          description: Datenqualitätsbewertung (0-1)
        verificationStatus:
          type: string
          enum: [unverified, verified, outdated]

    # IDENTIFICATION SCHEMAS
...
    CustomerDataRequest:
      type: object
      required:
        - sharedCustomerHash
        - requestedModules
        - consentToken
      properties:
        sharedCustomerHash:
          type: string
          description: SHA-256 Hash des Kunden
        requestedModules:
          type: array
          items:
            type: string
            enum: [basisdaten_module, erweiterte_daten_module, identifikation_module, background_checks_module]
          description: Angeforderte Datenbausteine
        consentToken:
          type: string
          description: JWT-Token mit Consent-Nachweis
        purpose:
          type: string
          enum: [kundenbeziehungseroffnung, re_identification, compliance_update]
    CustomerDataResponse:
      type: object
      properties:
        availableModules:
          type: array
          items:
            type: string
        moduleData:
          type: object
          description: Bereitgestellte Datenbausteine
        verificationStatus:
          type: object
        lastUpdated:
          type: string
          format: date-time
...